% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEi-package.R
\docType{package}
\name{TEi-package}
\alias{TEi}
\alias{TEi-package}
\title{TEi: transposable element insertion detection}
\description{
TEi provides a maintained short-read soft-clip workflow and an experimental
long-read workflow for Oxford Nanopore and PacBio genome alignments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YeehanXiao/TE_insert}
  \item \url{https://github.com/tchen-tt/TEi}
  \item Report bugs at \url{https://github.com/YeehanXiao/TE_insert/issues}
}

}
\author{
\strong{Maintainer}: Yihan Xiao \email{1653008@tongji.edu.cn}

Authors:
\itemize{
  \item Yihan Xiao \email{1653008@tongji.edu.cn}
  \item Tao Chen \email{taochenhm@gmail.com}
}

}
\keyword{internal}
